move : 🚚 컴포넌트 해당 페이지 별로 폴더화
