+++ week4/assignment/로그인 & 회원가입 👤/src/components/signup/DoubleCheck.jsx
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { Sign } from "../styles/commonStyle";
+import { Sign } from "../../styles/commonStyle";
 
 const DoubleCheck = ({ id, isAvailable, setIsAvailable }) => {
   const checkIdExist = async () => {
