+++ week4/assignment/로그인 & 회원가입 👤/src/components/signup/SingUp.jsx
@@ -1,8 +1,8 @@
 import { useNavigate } from "react-router-dom";
-import { Sign } from "../styles/commonStyle";
+import { Sign } from "../../styles/commonStyle";
 import axios from "axios";
 
-import { INPUT } from "../constants/input";
+import { INPUT } from "../../constants/input";
 
 const SignUp = ({ isInputsFilled, info }) => {
   const requestData = {
