+++ week4/assignment/로그인 & 회원가입 👤/src/pages/SignUpPage.jsx
@@ -1,4 +1,4 @@
-import SignForm from "../components/SignForm";
+import SignForm from "../components/common/SignForm";
 import { PAGE } from "../constants/page";
 
 const SignUpPage = () => {
